add modbus sensors support
