+++ MANIFEST.in
@@ -2,4 +2,5 @@ include README.md
 include requirements.txt
 include LICENSE
 recursive-include boneio *.yaml
+recursive-include boneio *.json
 recursive-include boneio *.ttf
