+++ boneio/const.py
@@ -35,7 +35,7 @@ UART = "uart"
 RX = "rx"
 TX = "tx"
 RESTORE_STATE = "restore_state"
-
+MODEL = "model"
 UARTS = {
     "uart1": {ID: "/dev/ttyS1", TX: "P9.24", RX: "P9.26"},
     "uart4": {ID: "/dev/ttyS4", TX: "P9.13", RX: "P9.11"},
@@ -104,7 +104,6 @@ BINARY_SENSOR = "binary_sensor"
 LM75 = "lm75"
 MCP_TEMP_9808 = "mcp9808"
 INPUT_SENSOR = "inputsensor"
-SDM630 = "sdm630"
 
 BASE = "base"
 LENGTH = "length"
