+++ boneio/sensor/modbus/sdm120.json
@@ -1,0 +1,122 @@
+{
+    "model": "SDM120",
+    "registers_base": [
+        {
+            "base": 0,
+            "length": 34,
+            "registers": [
+                {
+                    "name": "Voltage",
+                    "address": 0,
+                    "unit_of_measurement": "V",
+                    "state_class": "measurement",
+                    "device_class": "voltage",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "Current",
+                    "address": 6,
+                    "unit_of_measurement": "A",
+                    "state_class": "measurement",
+                    "device_class": "current",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "Power",
+                    "address": 12,
+                    "unit_of_measurement": "W",
+                    "state_class": "measurement",
+                    "device_class": "power",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "Power_VA",
+                    "address": 18,
+                    "unit_of_measurement": "VA",
+                    "state_class": "measurement",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "Power_VAr",
+                    "address": 24,
+                    "unit_of_measurement": "var",
+                    "state_class": "measurement",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "Power_Factor",
+                    "address": 30,
+                    "unit_of_measurement": "%",
+                    "state_class": "measurement",
+                    "return_type": "float32"
+                }
+            ]
+
+        },
+        {
+            "base": 70,
+            "length": 12,
+            "registers":[
+                {
+                    "name": "Frequency",
+                    "address": 70,
+                    "unit_of_measurement": "Hz",
+                    "state_class": "measurement",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "Import_Energy_kWh_Total",
+                    "address": 72,
+                    "unit_of_measurement": "kWh",
+                    "state_class": "total_increasing",
+                    "device_class": "energy",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "Export_Energy_kWh_Total",
+                    "address": 74,
+                    "unit_of_measurement": "kWh",
+                    "state_class": "total_increasing",
+                    "device_class": "energy",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "Import_Energy_kVArh_Total",
+                    "address": 76,
+                    "unit_of_measurement": "kvArh",
+                    "state_class": "total_increasing",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "Export_Energy_kVArh_Total",
+                    "address": 78,
+                    "unit_of_measurement": "kvArh",
+                    "state_class": "total_increasing",
+                    "return_type": "float32"
+                }
+            ]
+        },
+        {
+            "base": 342,
+            "length": 4,
+            "registers": [
+                {
+                    "name": "kWh_Total",
+                    "address": 342,
+                    "unit_of_measurement": "kWh",
+                    "state_class": "total_increasing",
+                    "device_class": "energy",
+                    "return_type": "float32"
+                },
+                {
+                    "name": "kvarh_Total",
+                    "address": 344,
+                    "unit_of_measurement": "kWh",
+                    "state_class": "total_increasing",
+                    "device_class": "energy",
+                    "return_type": "float32"
+                }
+            ]
+        }
+    ]
+}
